move Tag to component folder
