+++ website/src/components/Tag/styles.module.css
@@ -1,0 +1,19 @@
+.tag.required {
+  background-color: #e75471;
+}
+
+.tag.dependency {
+  background-color: #ec8f17;
+}
+
+.tag.id {
+  background-color: #00b3d3;
+}
+
+.tag.core {
+  background-color: #00c79d;
+}
+
+.tag.extended {
+  background-color: #a1a4b1;
+}
